requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiohttp==3.8.6
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
pandas==2.1.3
pyarrow==14.0.1
pytest==7.4.3
pytest-asyncio==0.21.1